Dropping columns process is separated depending on whether they do it with name or if they do it with index
